refactor: move db initialisation to root layout so it happens on every page
