+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import "../styles/globals.css";
 import type { Metadata, Viewport } from "next";
 import type { ReactNode } from "react";
 import Navbar from "@/ui/layout/Navbar/Navbar";
+import DatabaseInitializer from "@/lib/db/DatabaseInitialiser";
 
 const APP_NAME = "Things We Do";
 const APP_DEFAULT_TITLE = "Things We Do";
@@ -52,6 +53,7 @@ export default function RootLayout({ children }: { children: ReactNode }) {
     <html lang="en" dir="ltr">
       <head />
       <body>
+        <DatabaseInitializer />
         <main className="pb-24">{children}</main>
         <Navbar />
       </body>
